backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
python-dotenv
fastapi-mail
reportlab
psycopg2-binary
